Pull out a configuration library.
This has been done with minimal effort.

Further work is required to minimize its surface area.
