+++ crates/configuration/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "ndc-postgres-configuration"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+query-engine-metadata = { path = "../query-engine/metadata" }
+query-engine-sql = { path = "../query-engine/sql" }
+
+ndc-sdk = { git = "https://github.com/hasura/ndc-hub.git", rev = "3b6c480" }
+
+schemars = { version = "0.8.16", features = ["smol_str", "preserve_order"] }
+serde = "1.0.196"
+serde_json = { version = "1.0.113", features = ["raw_value"] }
+sqlx = { version = "0.7.3", features = [ "json", "postgres", "runtime-tokio-rustls" ] }
+tracing = "0.1.40"
