+++ crates/configuration/src/introspection-notes.md
