+++ crates/configuration/src/version1.sql
