+++ crates/configuration/src/version2.sql
