+++ crates/connectors/ndc-postgres/Cargo.toml
@@ -15,17 +15,17 @@ name = "ndc-postgres"
 path = "bin/main.rs"
 
 [dependencies]
-ndc-sdk = { git = "https://github.com/hasura/ndc-hub.git", rev = "3b6c480" }
+ndc-postgres-configuration = { path = "../../configuration" }
 query-engine-execution = { path = "../../query-engine/execution" }
 query-engine-metadata = { path = "../../query-engine/metadata" }
 query-engine-sql = { path = "../../query-engine/sql" }
 query-engine-translation = { path = "../../query-engine/translation" }
 
+ndc-sdk = { git = "https://github.com/hasura/ndc-hub.git", rev = "3b6c480" }
+
 async-trait = "0.1.77"
 percent-encoding = "2.3.1"
 prometheus = "0.13.3"
-schemars = { version = "0.8.16", features = ["smol_str", "preserve_order"] }
-serde = "1.0.196"
 serde_json = { version = "1.0.113", features = ["raw_value"] }
 sqlx = { version = "0.7.3", features = [ "json", "postgres", "runtime-tokio-rustls" ] }
 thiserror = "1.0"
