+++ crates/connectors/ndc-postgres/src/explain.rs
@@ -12,7 +12,8 @@ use ndc_sdk::models;
 use query_engine_sql::sql;
 use query_engine_translation::translation;
 
-use super::configuration;
+use ndc_postgres_configuration as configuration;
+
 use super::state;
 
 /// Explain a query by creating an execution plan
