+++ crates/connectors/ndc-postgres/src/lib.rs
@@ -1,7 +1,6 @@
 //! A Hasura v3 PostgreSQL Native Data Connector.
 
 pub mod capabilities;
-pub mod configuration;
 pub mod connector;
 pub mod explain;
 pub mod health;
