+++ crates/documentation/openapi/src/generator.rs
@@ -1,5 +1,6 @@
-use ndc_postgres::configuration::RawConfiguration;
 use schemars::{gen::SchemaSettings, schema::RootSchema};
+
+use ndc_postgres_configuration::RawConfiguration;
 
 pub fn generate_schema() -> RootSchema {
     SchemaSettings::openapi3()
