+++ crates/tests/databases-tests/Cargo.toml
@@ -22,6 +22,7 @@ postgres = []
 [dependencies]
 openapi-generator = { path = "../../documentation/openapi" }
 ndc-postgres = { path = "../../connectors/ndc-postgres" }
+ndc-postgres-configuration = { path = "../../configuration" }
 ndc-test = { git = "https://github.com/hasura/ndc-spec.git", tag = "v0.1.0-rc.13" }
 tests-common = { path = "../tests-common" }
