+++ crates/tests/tests-common/src/common_tests/configuration_v1_tests.rs
@@ -1,8 +1,9 @@
 use std::fs;
 use std::path::Path;
 
-use ndc_postgres::configuration::version1;
 use similar_asserts::assert_eq;
+
+use ndc_postgres_configuration::version1;
 
 use crate::ndc_metadata::helpers::get_path_from_project_root;
 use crate::schemas::check_value_conforms_to_schema;
